# server/internal/app/testdata/probe/openapi.yaml
openapi: 3.1.0
info: { title: The probe module's contract, version: '1' }
paths:
...
  /me:
    get: { operationId: getMe, responses: { '200': { description: The user }, default: { $ref: '#/components/responses/Problem' } } }
    patch: { operationId: patchMe, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '200': { description: Updated }, default: { $ref: '#/components/responses/Problem' } } }
  /me/avatar:
    put: { operationId: putMeAvatar, requestBody: { $ref: '#/components/requestBodies/Upload' }, responses: { '200': { description: The user }, default: { $ref: '#/components/responses/Problem' } } }
  /me/sessions:
    get: { operationId: getMeSessions, responses: { '200': { description: Sessions }, default: { $ref: '#/components/responses/Problem' } } }
    delete: { operationId: deleteMeSessions, responses: { '204': { description: Done }, default: { $ref: '#/components/responses/Problem' } } }
  /me/sessions/{session_id}:
    parameters: [{ name: session_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    delete: { operationId: deleteMeSessionsBySessionId, responses: { '204': { description: Revoked }, default: { $ref: '#/components/responses/Problem' } } }
  # The household routes the router always serves (item 10), declared as loosely as the tests need.
...
  /me/consents:
    get: { operationId: getMeConsents, responses: { '200': { description: Consents }, default: { $ref: '#/components/responses/Problem' } } }
    put: { operationId: putMeConsents, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '200': { description: Updated }, default: { $ref: '#/components/responses/Problem' } } }
...
  /me/diagnostics:
    post: { operationId: postMeDiagnostics, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '201': { description: Attached }, default: { $ref: '#/components/responses/Problem' } } }
...
components:
...
  requestBodies:
    Any: { required: true, content: { application/json: { schema: { type: object } } } }
    Upload: { required: true, content: { multipart/form-data: { schema: { type: object } } } }
  responses:
    Problem:
      description: A problem.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Uuid:
      type: string
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
...
    ProblemCode: { type: string }
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { $ref: '#/components/schemas/ProblemCode' }
        request_id: { type: string }
...